fix example bug

+++ example/mxnet/train_gluon_mnist_byteps_gc.py
@@ -148,7 +148,8 @@ model.cast(args.dtype)
 # Initialize parameters
 model.initialize(mx.init.MSRAPrelu(), ctx=context)
 # if bps.rank() == 0:
-model.summary(nd.ones((1, 1, 28, 28), ctx=mx.gpu(bps.local_rank())))
+model.summary(nd.ones((1, 1, 28, 28), ctx=mx.gpu(
+    bps.local_rank()).astype(args.dtype, copy=False)))
 model.hybridize()
 
 params = model.collect_params()
